+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 ### Fixed
 - Passing user id integer instead of user element when creating a subscription. ([#12](https://github.com/craftcms/commerce-stripe/commit/61090bedfa9efe40f23c57a77c25cb3e50d50438))
 
+### Added
+- Subscription Form model which supports coupon, prorate, and metadata [Subscription Arguments](https://stripe.com/docs/api#create_subscription) ([#14](https://github.com/craftcms/commerce-stripe/issues/14))  
+
 ## 1.0.2
 
 - Added support for card tokens, not just payment sources. ([#9](https://github.com/craftcms/commerce-stripe/issues/9))